[I25-171] feat : 모듈화
